#105 - re-use character profile image component in CreateMessage component
