+++ client/src/components/Games/View/components/GameMessages/GameMessages.styl
@@ -8,11 +8,6 @@
       .column
         margin-bottom 0 !important
 
-      .primary-attributes, .secondary-attributes
-        line-height 1rem
-      .secondary-attributes
-        font-size 12px
-
     .row.slim
       padding-top 0
       padding-bottom 0
@@ -20,27 +15,6 @@
       .column
         margin-top .5rem !important
         margin-bottom .5rem  !important
-
-    .character-name, .user-name
-      font-size 24px
-      font-weight bolder
-
-    .character-name
-      padding-bottom 2px
-
-    .secondary-attributes
-      .attribute
-        display inline-block
-        margin-right 8px
-
-        &:after
-          content: "\007C"
-          margin-left 8px
-
-        &:last-child
-          &:after
-            content: ""
-            margin-left 0
 
     .column-info
       font-size 10px
